inspector wired up both ways; split inspector components out; text input responding to mutate; MCQ partially mutating
