+++ assets/src/apps/delivery/components/PartComponent.tsx
@@ -31,7 +31,7 @@ const WebComponent: React.FC<any> = (props) => {
     ];
     const notifications = notificationsHandled.map((notificationType: NotificationType) => {
       const handler = (e: any) => {
-        console.log(`${notificationType.toString()} notification handled [PC : ${props.id}]`, e);
+        /* console.log(`${notificationType.toString()} notification handled [PC : ${props.id}]`, e); */
         const el = ref.current;
         if (el) {
           if (el.notify) {
